Refactor run functions into run.go
